+++ screens/SignInScreen.js
@@ -1,6 +1,6 @@
 import React from "react";
 
-import SignInTabView from '../components/SignInTabView';
+import SignInTabView from '../components/SignIn/SignInTabView';
 
 const SignInScreen = ({ navigation }) => {
   return (
